requirements.txt
langchain
langchain-community
langchain-deepseek
python-dotenv
